move SwiftXPC+JSON to JSONOutput
